+++ boilerplate/web_modules/Layout/index.js
@@ -11,11 +11,11 @@ export default class Layout extends Component {
 
   static propTypes = {
     children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
-  }
+  };
 
   static contextTypes = {
     metadata: PropTypes.object.isRequired,
-  }
+  };
 
   render() {
     const {