+++ boilerplate/web_modules/Page/index.js
@@ -15,11 +15,11 @@ export default class Page extends Component {
     __url: PropTypes.string.isRequired,
     head: PropTypes.object.isRequired,
     body: PropTypes.string.isRequired,
-  }
+  };
 
   static contextTypes = {
     metadata: PropTypes.object.isRequired,
-  }
+  };
 
   render() {
     const {