+++ docs/web_modules/Layout/index.js
@@ -13,11 +13,11 @@ export default class Layout extends Component {
   static propTypes = {
     children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
     params: PropTypes.object,
-  }
+  };
 
   static contextTypes = {
     metadata: PropTypes.object.isRequired,
-  }
+  };
 
   render() {
     const {